requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.3
idna==3.6
multidict==6.0.4
paho-mqtt==1.6.1
python-dateutil==2.8.2
pyTibber==0.28.2
six==1.16.0
sniffio==1.3.0
tzdata==2023.4
websockets==12.0
yarl==1.9.4
